Divided classes into header and source files
