+++ colour_detection/.gitignore
@@ -3,3 +3,4 @@ CMakeFiles/*
 cmake_install.cmake
 Makefile
 CMakeCache.txt
+.vscode/
