+++ colour_detection/CMakeLists.txt
@@ -2,5 +2,5 @@ cmake_minimum_required(VERSION 2.8)
 project( ColourDetection )
 find_package( OpenCV REQUIRED )
 include_directories( ${OpenCV_INCLUDE_DIRS} )
-add_executable( ColourDetection ColourDetection.cpp )
+add_executable( ColourDetection ColourDetection.cpp src/color_detection.cpp src/init.cpp )
 target_link_libraries( ColourDetection ${OpenCV_LIBS} )
